// src/Pages/SportDetails/MatchBet/types.ts
// types.ts
export interface Breadcrumb {
    name: string;
    link: string;
    class: string;
  }
  
  export interface Bet {
    sr: number;
    rate: string;
    amount: number;
    mode: string;
    team: string;
    client: string;
    agent: string;
    dateTime: string;
    shaheenHunters: number;
    kabulZalmi: number;
  }
  